+++ ActivityEvents.cs
@@ -29,5 +29,6 @@
         public const string Text = nameof(Text);
         public const string Ping = nameof(Ping);
         public const string Inventory = nameof(Inventory);
+        public const string TrinketActivated = nameof(TrinketActivated);
     }
 }